Moved decomposing logic into separate file
